requirement.txt
streamlit
PyPDF2
google-generativeai
python-dotenv
langchain
pdfplumber
langchain_community
langchain-groq
mysql-connector-python
openai==0.28
groq
huggingface-hub
transformers
torch
langchain_core
transformers
huggingface
sentencepiece
langchain-huggingface
vertexai
sqlalchemy